refactor for frontend
프론트엔드를 위한 약간의 코드 수정
